# gateway/cmd/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Volha gateway REST API",
        "contact": {
            "name": "Alex \"bustard\" Provor",
            "url": "https://breezy.su",
            "email": "help@breezy.su"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/files/delete": {
            "delete": {
                "description": "Удаляет файл. Требуется передовать только имя. Пример: example.gif",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Удалить файл",
                "parameters": [
                    {
                        "type": "string",
                        "description": "название файла",
                        "name": "title",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "файл успешно удалён",
                        "schema": {
                            "$ref": "#/definitions/views.SWGSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Неверное название",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/files/upload": {
            "post": {
                "description": "Сохраняет изображение в папке сервера images/. В поле \"img\" передается файл с расширениями jpg, png, webp, gif",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Загрузить изображение",
                "responses": {
                    "200": {
                        "description": "Цвет успешно создан",
                        "schema": {
                            "$ref": "#/definitions/views.SWGFileUploadResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат данных",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/material/create": {
            "post": {
                "description": "Добавляет новый материал",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "material"
                ],
                "summary": "Создать материал",
                "parameters": [
                    {
                        "description": "Новый материал",
                        "name": "material",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/views.Material"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Материал успешно создан",
                        "schema": {
                            "$ref": "#/definitions/views.SWGSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат данных",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Ошибка взаимодействия с сервисом",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/material/delete": {
            "delete": {
                "description": "Удаляет материал по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "material"
                ],
                "summary": "Удалить материал",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID материала",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Материал успешно удалён",
                        "schema": {
                            "$ref": "#/definitions/views.SWGSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Ошибка взаимодействия с сервисом",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/material/getall": {
            "get": {
                "description": "Возвращает список всех материалов",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "material"
                ],
                "summary": "Получить все материалы",
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "schema": {
                            "$ref": "#/definitions/views.SWGMaterialListResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Ошибка взаимодействия с сервисом",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/material/update": {
            "put": {
                "description": "Обновляет данные существующего материала",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "material"
                ],
                "summary": "Обновить материал",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID материала",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Обновлённые данные материала",
                        "name": "material",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/views.Material"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Материал успешно обновлён",
                        "schema": {
                            "$ref": "#/definitions/views.SWGSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный ID или данные",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка на сервере",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Ошибка взаимодействия с сервисом",
                        "schema": {
                            "$ref": "#/definitions/views.SWGErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "views.Material": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "views.SWGErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "something went wrong"
                }
            }
        },
        "views.SWGFileUploadResponse": {
            "type": "object",
            "properties": {
                "mime": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "views.SWGMaterialListResponse": {
            "type": "object",
            "properties": {
                "materials": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/views.Material"
                    }
                }
            }
        },
...
        "views.SWGSuccessResponse": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string",
                    "example": "operation successful"
                }
            }
        }
    }
}